requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.2.0
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
discord.py==1.7.3
idna==3.2
multidict==5.2.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.4.0
praw==7.4.0
prawcore==2.3.0
python-dateutil==2.8.2
PyYAML==5.4.1
redis==3.5.3
regex==2021.9.30
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==1.2.1
tweepy==4.10.0
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.6.3
